backend/requirements.txt
anyio==3.6.1
asgiref==3.8.1
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
crashtest==0.3.1
distlib==0.3.6
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
dulwich==0.20.46
fastapi==0.85.0
filelock==3.8.0
h11==0.14.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.1
poetry-core==1.2.0
poetry-plugin-export==1.1.1
ptyprocess==0.7.0
pyaml==23.9.7
pycparser==2.21
pydantic==1.10.2
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
PyYAML==6.0.1
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.4
starlette==0.20.4
tomlkit==0.11.5
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.16.5
webencodings==0.5.1
